Enhance handling of failed InsertRequest calls
If the call to InsertRequest fails, we still have a successfully created
local request (referred to as a submission). This commit adds the reason
the InsertRequest call failed to the Staff Notes field on the local
request.

We also pass the failure reason back in the template response, this is
not used at the OPAC end, but will be used in the staff UI